# Add ClientTracer from httptrace for detailed connection info