Provide detailed connection information to help figure
out where a request is spending most of its time.

## Demo
Given this code below
```go
package main

import (
	"context"
	"io"
	"io/ioutil"
	"log"
	"net/http"

	"go.opencensus.io/exporter/stackdriver"
	"go.opencensus.io/plugin/ochttp"
	"go.opencensus.io/trace"
)

func main() {
	sd, err := stackdriver.NewExporter(stackdriver.Options{ProjectID: "census-demos"})
	if err != nil {
		log.Fatalf("Failed to create Stackdriver exporter: %v", err)
	}
	trace.RegisterExporter(sd)
	defer sd.Flush()

	trace.ApplyConfig(trace.Config{DefaultSampler: trace.AlwaysSample()})

	ctx, span := trace.StartSpan(context.Background(), "Fetching URL")
	defer span.End()

	tr := &ochttp.Transport{}
	client := &http.Client{Transport: tr}
	req, _ := http.NewRequest("GET", "https://opencensus.io", nil)
	req = req.WithContext(ctx)
	res, err := client.Do(req)
	if err != nil {
		log.Fatalf("Failed to GET: %v", err)
	}
	_, _ = io.Copy(ioutil.Discard, res.Body)
	_ = res.Body.Close()
}
```

and running it 
```shell
GOOGLE_APPLICATION_CREDENTIALS=~/Downloads/census-demos-237a8e1e41df.json go run main.go 
```

### Before
<img width="1197" alt="screen shot 2018-05-04 at 10 47 22 pm" src="https://user-images.githubusercontent.com/4898263/39660121-44a141f2-4fed-11e8-9b90-e3e16e884b16.png">

### After
<img width="1205" alt="screen shot 2018-05-04 at 10 48 02 pm" src="https://user-images.githubusercontent.com/4898263/39660119-4015f13c-4fed-11e8-9bd5-0c297df95dd9.png">