@basvanbeek 2018-05-06
I would not automatically enable this for all http client spans. This is expensive enough to be off by default and enabled by config when needed.