@basvanbeek 2018-05-06
@ramonza I disagree... this is a choice we should allow the consumer to make. We can argue about it being important enough with acceptable overhead for it to be on by default but it should be adjustable by the end user.

Personally I think OC overhead should be minimized by default and additional details should be enabled when needed. No need in having customers incur higher cost or needing to add resources because we're being overly verbose by default.