@ramonza 2018-05-06
@basvanbeek are you worried about resource consumption (CPU, memory) or trace verbosity? 

If resource consumption, then doesn't sampling handle this already? The user can decrease the sampling if the overhead is too high. It would be good to have some benchmarks here to quantify it. I would prefer not to add configuration complexity if the overhead isn't large compared to the overhead we already incur anyway if a trace is sampled (collecting and exporting).

As for verbosity of traces an alternative would be to log these as events/annotations rather than sub-spans.