@basvanbeek 2018-05-06
This will be a very long winded conversation on why we should limit amount of data by default which I do not have time for at this moment to explain.

But just to name a few: 
- lowering sampling rate especially with unweighted endpoints means that valuable information on certain endpoints might be dropped over relatively unexciting high volume endpoints due to chance of being sampled going drastically down.
- lots of times trace investigation is about finding issues in micro service interop, parallelism vs serialism, failures down the tree and information on connection establishment is not really high on the list of things to see if most service to service calls are local.
- storage / retrieval costs on various backends we wish to support might be a bottleneck.
- lots of users start with high sampling rates to simply find out how their system behaves. When they reach high volumes they start to sample, but do this relatively late. forcing this decision to come sooner by over verbosity might not be that great of a story to users.
- OC instrumentation for Go is already on the heavy side with respect to time used on the critical path, no need to make this worse when not explicitly called for by consumer.

Me as a consumer would really dislike for you as an instrumentation engineer to force these decisions on me. I'm very well capable of configuring the items I want / don't want as long as you give me that option. Don't expect you know all the use cases and configuration needs as I'm convinced I don't know either.

As said it is debatable if config for this option should be on or off by default but I'll fight for the option for it to be setable.