@ramonza 2018-05-06
@odeke-em what are your thoughts on this? Did you do any benchmarks? Might be good to add benchmarks since the issue has come up now.

I do think that if we add options for performance reasons, we need to support this claim with benchmarks showing the effect of the option. 

Thinking more about it, I would actually double down on the suggestion to use annotations rather than sub-spans. This would reduce the verbosity and probably(?) the resource cost as well. @basvanbeek do you have thoughts on this?