@odeke-em 2018-05-09
Thank you @ramonza and @basvanbeek for the reviews and for the discourse.
In deed, as @basvanbeek advocated against, this change would be very latency heavy as the benchmarks show
```go
var sink *http.Response

func BenchmarkRoundTrip(b *testing.B) {
        b.ReportAllocs()
        // We need a functional server to test out client tracing
        cst := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
                _, _ = ioutil.ReadAll(r.Body)
                _ = r.Body.Close()
                w.Write([]byte("Hello, there"))
        }))
        defer cst.Close()

        client := &http.Client{Transport: new(Transport)}
        for i := 0; i < b.N; i++ {
                res, err := client.Get(cst.URL)
                if err != nil {
                        b.Fatalf("Failed to invoke get: %v", err)
                }
                _ = res.Body.Close()
                sink = res
        }
        if sink == nil {
                b.Fatalf("Did not run benchmark!")
        }
}
```

### After
```shell
BenchmarkRoundTrip-4   	    5000	    380716 ns/op	   21315 B/op	     208 allocs/op
```

### Before
```shell
BenchmarkRoundTrip-4   	    2000	    816372 ns/op	   20333 B/op	     175 allocs/op
```

With annotations added, the number of allocs increases by a bit but not as badly as with subspans
```shell
BenchmarkRoundTrip-4   	    3000	    377179 ns/op	   20836 B/op	     195 allocs/op
```

My use case was investigating which parts of a network bound request were pathologically slow as the rest of the part of my microservices returned in at most 5ms but the network cost me like 400ms or so. 

@basvanbeek and you @ramonza both raise great points; I think the extent of the allocations and latency created don't give this change merit. However, an advanced user that really would like to trace through their request perhaps might be able to attach a clientTracer to their transport.

The annotations work well and are less heavier but perhaps not every user needs as much detail as I did need, in my very niche case?